HBASE-10116 Addendum: fix findbugs warning
git-svn-id: https://svn.apache.org/repos/asf/hbase/trunk@1550522 13f79535-47bb-0310-9956-ffa450edef68

+++ hbase-server/src/main/java/org/apache/hadoop/hbase/io/hfile/slab/SlabCache.java
@@ -179,7 +179,7 @@ public class SlabCache implements SlabItemActionWatcher, BlockCache, HeapSize {
 
   private void addSlab(int blockSize, int numBlocks) {
     LOG.info("Creating a slab of blockSize " + blockSize + " with " + numBlocks
-        + " blocks, " + StringUtils.humanReadableInt(blockSize * numBlocks) + "bytes.");
+        + " blocks, " + StringUtils.humanReadableInt(blockSize * (long) numBlocks) + "bytes.");
     sizer.put(blockSize, new SingleSizeCache(blockSize, numBlocks, this));
   }
